requirements/requirements.txt
arrow==0.14.4
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.8
djangorestframework==3.11.0
factory-boy==2.8.1
Faker==4.1.1
gunicorn==20.0.4
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
wcwidth==0.2.5
whitenoise==5.1.0
psycopg2==2.8.5
